Use routing and template app.component
