+++ src/app/app.component.html
@@ -1,43 +1,6 @@
-<!--The content below is only a placeholder and can be replaced.-->
-<div style="text-align:center">
-	<h1>
-		Welcome to Mixed Baggle!
-	</h1>
+<div id="page">
+    <app-header></app-header>
+    <div style="min-height: 700px;">
+        <router-outlet></router-outlet>
+    </div>
 </div>
-<div>
-	<h3>Customize your board!</h3>
-	<settings [settings]="settings"></settings>
-	<div>
-		<input type="button" value="Generate Board!" (click)="generateBoard()">
-	</div>
-</div>
-<div>
-	<div class="col-sm-5">
-		<game-board [dice]="shakenDice"></game-board>
-	</div>
-	<div class="col-sm-5">
-		<input type="text" [(ngModel)]="enteredText" (ngModelChange)="showOnBoard()" (keyup.enter)="onKey($event)">
-		<div>
-			<div>Points: {{totalPoints}}</div>
-			<div *ngFor="let guessedWord of toArray(guessedWords)">
-				{{ guessedWord.key }} - {{ guessedWord.value }}
-			</div>
-		</div>
-	</div>
-</div>
-
-<div>
-	<div>Found {{wordCount}} words in {{timeTaken}} milliseconds.</div>
-	<div>
-		<fa-icon [icon]="faAngleDown" *ngIf="showFoundWords" (click)="showFoundWords = !showFoundWords"></fa-icon>
-		<fa-icon [icon]="faAngleUp" *ngIf="!showFoundWords" (click)="showFoundWords = !showFoundWords"></fa-icon>
-	</div>
-	<div *ngIf="showFoundWords">
-		<div *ngFor="let word of toArray(wordList)">
-			{{word.key | json}} --> {{word.value}}
-		</div>
-	</div>
-</div>
-
-
-<router-outlet></router-outlet>
