+++ src/app/app.component.scss
@@ -1,0 +1,5 @@
+#page {
+    padding-top: 30px;
+    min-height: 820px;
+    background: rgba(50,140,193, 0.7);
+}
