+++ src/app/header/header.component.html
@@ -1,0 +1,5 @@
+<div style="text-align:center">
+	<h1>
+		Welcome to Mixed Baggle!
+	</h1>
+</div>
